% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rds_files}
\alias{get_rds_files}
\title{Get Seurat Objects from a Folder}
\usage{
get_rds_files(folder, base_folder)
}
\arguments{
\item{folder}{The path to the folder to search in.}

\item{base_folder}{The base data directory path.}
}
\value{
A character vector of file names.
}
\description{
Lists RDS files in a given folder that may contain Seurat objects.
}
\keyword{internal}
