% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_sample}
\alias{load_sample}
\title{Load a Seurat Object}
\usage{
load_sample(sample_file, folder, base_folder)
}
\arguments{
\item{sample_file}{The name of the RDS file to load.}

\item{folder}{The folder containing the file.}

\item{base_folder}{The base data directory path.}
}
\value{
A Seurat object or NULL if loading fails.
}
\description{
Loads a Seurat object from an RDS file.
}
\keyword{internal}
